requirements.txt
# Core ML and Deep Learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Development and Analysis
jupyter>=1.0.0
notebook>=7.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
